+++ src/context/homeContext.ts
@@ -2,26 +2,18 @@ import { createContext } from 'react';
 
 // 상태 타입 정의
 export interface NavContextType {
-  isBoardOpened: boolean;
-  setIsBoardOpened: React.Dispatch<React.SetStateAction<boolean>>;
-  openModalBoard: () => void;
   inputBoardData: (sort: string, value: string | number) => void;
   writeBoard: () => void;
-  sortingBoards: (value: string) => void;
   handleBoardImg: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  sortingBoards: (value: string) => void;
   sortValues: string[];
-  // sendMessage: () => Promise<void>;
 }
 
 // Context 생성 및 초기값 설정
 export const navContext = createContext<NavContextType>({
-  isBoardOpened: false,
-  setIsBoardOpened: () => {},
-  openModalBoard: () => {},
   inputBoardData: () => {},
   writeBoard: () => {},
-  sortingBoards: () => {},
   handleBoardImg: () => {},
+  sortingBoards: () => {},
   sortValues: [],
-  // sendMessage: () => {},
 });
